fix: replace deprecated Json::Reader and fix type conversion warnings
Replace deprecated Jsoncpp API usage and fix type conversion warnings
to achieve clean compilation without warnings.

**Changes:**

1. **WechatCallbackController.cc** - Replace Json::Reader with CharReaderBuilder
   - Use modern CharReaderBuilder instead of deprecated Reader
   - Add proper error handling with string errors

2. **AlipaySandboxClient.cc** - Replace Json::Reader with CharReaderBuilder
   - Update JSON parsing to use modern Jsoncpp API
   - Maintain error handling with proper string conversion

3. **WechatPayClient.cc** - Fix size_t to int conversion warnings
   - Add explicit static_cast<int>() for size_t parameters
   - Fix nonce.size() conversion (line 229)
   - Fix aad.size() conversion (line 252)
   - Fix textLen conversion (line 266)

**Warnings Fixed:**
- ✅ C4996: 'Json::Reader::parse' deprecation warnings (3 instances)
- ✅ C4267: size_t to int conversion warnings (4 instances)

**Verification:**
- ✅ Clean compilation with zero warnings
- ✅ PayServer.exe builds successfully (6.6MB)
- ✅ Code uses modern Jsoncpp API best practices

**Impact:**
- Code now uses recommended Jsoncpp API
- Type-safe conversions prevent potential data loss
- Improved maintainability and future compatibility

+++ PayBackend/plugins/WechatPayClient.cc
@@ -226,7 +226,7 @@ bool decryptAesGcm(const std::string &ciphertextB64,
         return false;
     }
 
-    if (EVP_CIPHER_CTX_ctrl(ctx, EVP_CTRL_GCM_SET_IVLEN, nonce.size(),
+    if (EVP_CIPHER_CTX_ctrl(ctx, EVP_CTRL_GCM_SET_IVLEN, static_cast<int>(nonce.size()),
                             nullptr) != 1)
     {
         EVP_CIPHER_CTX_free(ctx);
@@ -249,7 +249,7 @@ bool decryptAesGcm(const std::string &ciphertextB64,
     {
         if (EVP_DecryptUpdate(ctx, nullptr, &outLen,
                               reinterpret_cast<const unsigned char *>(aad.data()),
-                              aad.size()) != 1)
+                              static_cast<int>(aad.size())) != 1)
         {
             EVP_CIPHER_CTX_free(ctx);
             error = "set aad failed";
@@ -263,7 +263,7 @@ bool decryptAesGcm(const std::string &ciphertextB64,
             reinterpret_cast<unsigned char *>(&plaintext[0]),
             &outLen,
             reinterpret_cast<const unsigned char *>(ciphertext.data()),
-            textLen) != 1)
+            static_cast<int>(textLen)) != 1)
     {
         EVP_CIPHER_CTX_free(ctx);
         error = "decrypt update failed";
